Move factory interface to public package
